split cgroups discovery to a separate file
