+++ Makefile.am
@@ -280,6 +280,8 @@ IDLEJITTER_PLUGIN_FILES = \
     $(NULL)
 
 CGROUPS_PLUGIN_FILES = \
+    collectors/cgroups.plugin/cgroup-internals.h \
+    collectors/cgroups.plugin/cgroup-discovery.c \
     collectors/cgroups.plugin/sys_fs_cgroup.c \
     collectors/cgroups.plugin/sys_fs_cgroup.h \
     $(NULL)
